Subset of WMA links up to e1..
Plus addressing long url lines, dt, apply->eval
